% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\docType{data}
\name{cov}
\alias{cov}
\title{Batch information  for samples in hnsc.edata.}
\format{A matrix of covariate(s) in columns.}
\usage{
cov
}
\description{
Batch information  for samples in hnsc.edata.
}
\keyword{datasets}
