% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{design}
\alias{design}
\title{Design matrix for samples in hnsc.edata.}
\format{A model matrix with 80 rows (i.e. samples) and 3 columns of tissue type and batch.}
\usage{
design
}
\description{
Design matrix for samples in hnsc.edata.
}
\keyword{datasets}
