% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnsc.edata.R
\docType{data}
\name{hnsc.edata}
\alias{hnsc.edata}
\title{lncRNA Fragments Per Killobase per Million (FPKM) in a head and neck squamous cell carcinomas (hnsc) study.}
\format{A data frame of lncRNA FPKM with 1000 rows (i.e. genes) and 80 columns (i.e. samples ).}
\usage{
hnsc.edata
}
\description{
lncRNA Fragments Per Killobase per Million (FPKM) in a head and neck squamous cell carcinomas (hnsc) study.
}
\keyword{datasets}
