% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.hnsc.match.edata.R
\docType{data}
\name{tcga.hnsc.match.edata}
\alias{tcga.hnsc.match.edata}
\title{TCGA head and neck squamous cell carcinomas (HNSC) lncRNA Reads Per Killobase per Million (RPKM).}
\format{A data frame of lncRNA RPKM with 1322 rows (i.e. genes) and 80 columns (i.e. samples ).}
\usage{
tcga.hnsc.match.edata
}
\description{
TCGA head and neck squamous cell carcinomas (HNSC) lncRNA Reads Per Killobase per Million (RPKM).
}
\keyword{datasets}
