% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.R
\docType{data}
\name{tissue}
\alias{tissue}
\title{Tissue type for samples in hnsc.edata.}
\format{A character vector of tissue type.}
\usage{
tissue
}
\description{
Tissue type for samples in hnsc.edata.
}
\keyword{datasets}
