% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIQML.LRT.R
\name{ZIQML.LRT}
\alias{ZIQML.LRT}
\title{Likelihood ratio test based on ZIQML.fit()}
\usage{
ZIQML.LRT(ZIQML.fit, coef = NULL, simulated.pvalue = F,
  permutation = 100)
}
\arguments{
\item{ZIQML.fit}{Object returned by ZIQML.fit()}

\item{coef}{An integer or vector indicating the coefficient(s) in design matrix to be tested. coef=1 is the intercept (i.e. baseline group effect),
and should not be tested.}

\item{simulated.pvalue}{If empirical p-values are computed, simulated.pvalue=TRUE. The default is FALSE.}

\item{permutation}{The number of permutations used in simulating pvalues. The defualt value is 100.}
}
\value{
\item{test.results}{Likelihood ratio test results with test statistics, p-value, FDR (or adjusted pvalues). 
       If simulated.pvalue=TRUE, test.results also includes simulated p-value and FDR.}  
       \item{Estimates}{Estimated group effect in object ZIQML.fit.}
}
\description{
ZIQML.LRT returns the likelihood ratio test results based on the object returned by ZIQML.fit().
}
\examples{

data('tcga.hnsc.match.edata','design')  
# 'tcga.hnsc.match.edata' contains RPKM of 1132 lncRNA genes and 80 samples. 
# 'design' is the design matrix for tumor vs normal tissue. 

fit.log=ZIQML.fit(edata=tcga.hnsc.match.edata,design.matrix=design,link='log') 
# Fit GLM by ZIQML with logarithmic link function 

LRT.results=ZIQML.LRT(fit.log,coef=2)  
# Likelihood ratio test on tumor vs normal, using observed p-values. 

# Include simulated p-values
fit.log.sim=ZIQML.fit(edata=tcga.hnsc.match.edata[1:10,],design.matrix=design,link='log') 
LRT.results.sim=ZIQML.LRT(fit.log.sim,coef=2,simulated.pvalue=TRUE,permutation=200)  



}
