% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\docType{data}
\name{design}
\alias{design}
\title{Design matrix for matched tumor-normal samples in tcga.hnsc.match.edata.}
\format{A model matrix with 80 rows (i.e. samples) and 2 columns (tumor vs normal).}
\usage{
design
}
\description{
Design matrix for matched tumor-normal samples in tcga.hnsc.match.edata.
}
\keyword{datasets}
