% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIQML.fit.R
\name{ZIQML.fit}
\alias{ZIQML.fit}
\title{Group effect estimate on lncRNA normalized counts}
\usage{
ZIQML.fit(edata, design.matrix, link = "log")
}
\arguments{
\item{edata}{Normalized counts matrix with genes in rows and samples in columns.}

\item{design.matrix}{Design matrix for groups, generated by model.matrix().}

\item{link}{Link function for the generalized linear model and likelihood function,either 'log' or 'identity'. The default is 'log'.}
}
\value{
\item{Estimates}{Estimated group effect on input genes by zero-inflated exponential quasi maximum likelihood (ZIQML) estimator.}
   \item{logLikelihood}{The value of zero-inflated quasi likelihood at estimated group effect.} 
   \item{edata}{The input gene expression matrix.} 
   \item{design.matrix}{The input design matrix.} 
   \item{link}{The specified link function.}
}
\description{
ZIQML.fit estimates the group effect on genes using zero-inflated exponential quasi likelihood.
}
\examples{

data('tcga.hnsc.match.edata','design') 
# 'tcga.hnsc.match.edata' contains RPKM of 1132 lncRNA genes and 80 samples 
# 'design' is the design matrix for tumor vs normal tissue.
                                       
fit.log=ZIQML.fit(edata=tcga.hnsc.match.edata,design.matrix=design,link='log') 
# Fit GLM by ZIQML with logarithmic link function 

fit.identity=ZIQML.fit(edata=tcga.hnsc.match.edata,design.matrix=design,link='identity') 
# Fit GLM by ZIQML with identity link function
}
